steps: Merge path.py with file.py
